requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
lxml>=5.2.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
scikit-learn>=1.4.0
pandas>=2.2.0
joblib>=1.4.0
